Enable FE proxying history reads to history service as default
