+++ config/dynamicconfig/development-cass.yaml
@@ -40,10 +40,6 @@ frontend.workerVersioningDataAPIs:
   - value: true
 frontend.workerVersioningWorkflowAPIs:
   - value: true
-frontend.accessHistoryFraction:
-  - value: 1.0
-frontend.adminDeleteAccessHistoryFraction:
-  - value: 1.0
 system.enableNexus:
   - value: true
 component.nexusoperations.callback.endpoint.template:
