+++ config/dynamicconfig/development-xdc.yaml
@@ -40,10 +40,6 @@ frontend.workerVersioningDataAPIs:
   - value: true
 frontend.workerVersioningWorkflowAPIs:
   - value: true
-frontend.accessHistoryFraction:
-  - value: 1.0
-frontend.adminDeleteAccessHistoryFraction:
-  - value: 1.0
 system.enableNexus:
   # enableNexus also controls transition history which can't be enabled now
   - value: false
