+++ tests/dynamicconfig.go
@@ -53,7 +53,6 @@ var (
 		dynamicconfig.NamespaceCacheRefreshInterval.Key():                       NamespaceCacheRefreshInterval,
 		dynamicconfig.FrontendEnableUpdateWorkflowExecution.Key():               true,
 		dynamicconfig.FrontendEnableUpdateWorkflowExecutionAsyncAccepted.Key():  true,
-		dynamicconfig.FrontendAccessHistoryFraction.Key():                       0.5,
 		dynamicconfig.ReplicationEnableUpdateWithNewTaskMerge.Key():             true,
 		dynamicconfig.ValidateUTF8SampleRPCRequest.Key():                        1.0,
 		dynamicconfig.ValidateUTF8SampleRPCResponse.Key():                       1.0,
